Added a suspicious case for policy change
Bog budget in the unit's currentcycle goes from 105 before policy
change to 10 after policy change when I expected it to stay at 105
despite the policy's max BOG changing from 365 to 270, but maybe
this is expected behavior.  The reason I think this might be an
issue is because a BOG budget of 10 is < overlap and I remember
that being an issue in the past.  This run won't throw an error
though so maybe it doesn't help.